django-ecommerce-master/requirements.txt
asgiref==3.8.1
autopep8==2.1.1
blinker==1.8.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2024.1
defusedxml==0.7.1
dj-database-url==2.2.0
Django==5.0.6
django-allauth==0.63.2
django-countries==7.6.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
Flask==3.0.3
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.1
pep8==1.7.1
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.1
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
requests==2.32.2
requests-oauthlib==2.0.0
runserver==1.0.4
sqlparse==0.5.0
stripe==9.8.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
